Move MainPanel to its own package
