+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/SetupPanel.java
@@ -23,7 +23,7 @@ import games.strategy.engine.data.PlayerID;
 import games.strategy.engine.framework.startup.launcher.ILauncher;
 import games.strategy.ui.SwingAction;
 
-abstract class SetupPanel extends JPanel implements ISetupPanel {
+public abstract class SetupPanel extends JPanel implements ISetupPanel {
   private static final long serialVersionUID = 4001323470187210773L;
   private static final String SET_ALL_DEFAULT_LABEL = "Default";
